appIconBar: Properly show running (and focused) app indicator
For some reason setting 'y_align' on the container is making it
ignore 'y_expand' (set to true) which in turn makes the
container too small to show the running app indicator (set via CSS
to a fixed position that considers the container will have the
same height as the panel).

Simply not setting 'y_align' makes things work as expected.

https://phabricator.endlessm.com/T30625