hotCorner: Add indicator menu to enable/disable hotcorner targets
Note that this setting controls whether hotcorner targets that are
triggerred by holding the pointer on one of the four corners of a screen
(depending on the settings) are enabled but it does not allow disabling
the action that is triggered by clicking on the hotcorner button.

Original code imported from Endless 3.8 gnome-shell.

https://phabricator.endlessm.com/T30445