Add specialized remote access indicator
The upstream impl hides the indicator on non wayland sessions
and given we enable X11 on Endless OS atm let's enable it
here. The original code was copied from gnome-shell 3.38 and
we create a new class instead of specializing the upstream one
to avoid some hacks that would be needed to handle object
destruction properly.

https://phabricator.endlessm.com/T30172