boxpointer: Calculate arrow position using minimum sizes
The function to calculate the final box pointer position uses the
natural size to determine whether to flip on the axis or not. In
the case of the date menu, this is problematic because it uses
StScrollView, which has a tiny minimum size and potentially large
natural sizes, a case where box pointer happily flips the axis
even when it should not.

Use minimum sizes to calculate final arrow position.

https://phabricator.endlessm.com/T30869