panel: Ensure indicator menus point to the right position
They currently assume the source actor is at the top, but
the panel extension moves all these indicators to the bottom,
which breaks the popup animation and tricks the inner box
pointer into thinking it should display below the panel.

Update indicator menu arrows to bottom when the panel extension
is active.

https://phabricator.endlessm.com/T30869